+++ src/admin.h
@@ -68,6 +68,9 @@ class AdminClient : public Connection {
   Baton DeleteRecords(rd_kafka_DeleteRecords_t** del_records,
                       size_t del_records_cnt, int operation_timeout_ms,
                       int timeout_ms, rd_kafka_event_t** event_response);
+  Baton DescribeTopics(rd_kafka_TopicCollection_t* topics,
+                       bool include_authorized_operations, int timeout_ms,
+                       rd_kafka_event_t** event_response);
 
  protected:
   static Nan::Persistent<v8::Function> constructor;
@@ -91,6 +94,7 @@ class AdminClient : public Connection {
   static NAN_METHOD(NodeDeleteGroups);
   static NAN_METHOD(NodeListConsumerGroupOffsets);
   static NAN_METHOD(NodeDeleteRecords);
+  static NAN_METHOD(NodeDescribeTopics);
 
   static NAN_METHOD(NodeConnect);
   static NAN_METHOD(NodeDisconnect);