+++ CHANGELOG.md
@@ -7,6 +7,7 @@ v0.5.0 is a limited availability feature release. It is supported for all usage.
 
 1. Add support for an Admin API to delete records.(#141).
 2. Fixes an issue with unresolved raced Promises leaking in the consumer (#151).
+3. Add support for an Admin API to describe topics.(#155).
 
 
 # confluent-kafka-javascript v0.4.0